refactor(work): move the scoreability projection into verify_scoreability
Keeps work_cmd/verification.py under the frozen 2380-line ratchet (now 2360) by moving the outcome_scoreability projection and its helpers into work_cmd/verify_scoreability.py. Behavior unchanged.

Receipts: 20260902-033155-work-verify-348d56, 20260902-033226-work-verify-dfa555.
